# services/api/src/routes/auth.ts
/**
 * Authentication routes
 * Handles login with Google/Apple, token refresh, and logout
 */

import type { FastifyInstance } from 'fastify';
import type { ZodTypeProvider } from 'fastify-type-provider-zod';
import { z } from 'zod';
import { eq, or } from 'drizzle-orm';
import { createPublicKey, createVerify, type JsonWebKey } from 'node:crypto';
import { db } from '../db/index.js';
import { refreshTokenRevocations, users } from '../db/schema.js';
import { config } from '../config.js';
import {
  generateAccessToken,
  type RefreshTokenPayload,
  generateRefreshToken,
  verifyRefreshToken,
  getAccessTokenExpiry,
} from '../plugins/auth.js';
import { getKarmaRank } from '../services/karma.js';
import { withGeneratedUniqueUsername } from '../utils/username.js';

// Validation schemas
const refreshSchema = z.object({
  refreshToken: z.string().min(1),
});

const logoutSchema = z.object({
  refreshToken: z.string().min(1).optional(),
});

interface AppleTokenHeader {
  alg: string;
  kid: string;
  typ?: string;
}

interface AppleTokenClaims {
  iss: string;
  aud: string | string[];
  exp: number;
  iat?: number;
  sub: string;
  email?: string;
  email_verified?: string | boolean;
}

function decodeJwtSegment<T>(token: string, index: number): T | null {
  const segment = token.split('.')[index];
  if (!segment) {
    return null;
  }

  try {
    return JSON.parse(Buffer.from(segment, 'base64url').toString('utf8')) as T;
  } catch {
    return null;
  }
}

async function fetchAppleSigningKey(kid: string) {
  const response = await fetch('https://appleid.apple.com/auth/keys', {
    signal: AbortSignal.timeout(5000),
  });
  if (!response.ok) {
    return null;
  }

  const body = (await response.json()) as {
    keys?: Array<JsonWebKey & { kid?: string }>;
  };
  const jwk = body.keys?.find((candidate) => candidate.kid === kid);
  if (!jwk) {
    return null;
  }

  return createPublicKey({
    key: jwk,
    format: 'jwk',
  });
}

async function verifyAppleJwt(idToken: string): Promise<AppleTokenClaims | null> {
  const header = decodeJwtSegment<AppleTokenHeader>(idToken, 0);
  const claims = decodeJwtSegment<AppleTokenClaims>(idToken, 1);

  if (!header || !claims || header.alg !== 'RS256' || !header.kid) {
    return null;
  }

  const expectedAudience = config.auth.appleClientId;
  const audienceMatches = Array.isArray(claims.aud)
    ? claims.aud.includes(expectedAudience)
    : claims.aud === expectedAudience;

  if (
    claims.iss !== 'https://appleid.apple.com' ||
    !audienceMatches ||
    claims.exp * 1000 <= Date.now() ||
    !claims.sub
  ) {
    return null;
  }

  const signingKey = await fetchAppleSigningKey(header.kid);
  if (!signingKey) {
    return null;
  }

  const [encodedHeader, encodedPayload, encodedSignature] = idToken.split('.');
  if (!encodedHeader || !encodedPayload || !encodedSignature) {
    return null;
  }

  const verifier = createVerify('RSA-SHA256');
  verifier.update(`${encodedHeader}.${encodedPayload}`);
  verifier.end();

  const valid = verifier.verify(
    signingKey,
    Buffer.from(encodedSignature, 'base64url'),
  );

  return valid ? claims : null;
}

/**
 * Validate Google ID token
 * In production, this would verify with Google's API
 * For development, we mock the validation
 */
async function validateGoogleToken(
  idToken: string
): Promise<{ email: string; googleId: string; name?: string } | null> {
  if (config.isDev === true) {
    // Mock validation for development
    // Token format: mock-google-{email}-{googleId}
    if (idToken.startsWith('mock-google-')) {
      const parts = idToken.split('-');
      if (parts.length >= 4) {
        return {
          email: parts[2] + '@gmail.com',
          googleId: parts[3],
          name: parts[2],
        };
      }
    }

    // For any token in dev mode, create a test user
    const timestamp = Date.now();
    return {
      email: `testuser${timestamp}@gmail.com`,
      googleId: `google-${timestamp}`,
      name: 'Test User',
    };
  }

  // Production: Verify with Google
  try {
    const response = await fetch(
      `https://oauth2.googleapis.com/tokeninfo?id_token=${idToken}`
    );

    if (!response.ok) {
      return null;
    }

    const data = (await response.json()) as {
      email: string;
      sub: string;
      name?: string;
      aud: string;
    };

    // Verify the audience matches our client ID
    if (data.aud !== config.auth.googleClientId) {
      return null;
    }

    return {
      email: data.email,
      googleId: data.sub,
      name: data.name,
    };
  } catch {
    return null;
  }
}

/**
 * Validate Apple ID token
 * In production, this would verify with Apple's API
 * For development, we mock the validation
 */
async function validateAppleToken(
  idToken: string
): Promise<{ email: string | null; appleId: string; name?: string } | null> {
  if (config.isDev === true) {
    // Mock validation for development
    if (idToken.startsWith('mock-apple-')) {
      const parts = idToken.split('-');
      if (parts.length >= 4) {
        return {
          email: parts[2] + '@privaterelay.appleid.com',
          appleId: parts[3],
          name: parts[2],
        };
      }
    }

    // For any token in dev mode, create a test user
    const timestamp = Date.now();
    return {
      email: `testuser${timestamp}@privaterelay.appleid.com`,
      appleId: `apple-${timestamp}`,
      name: 'Test User',
    };
  }

  try {
    const claims = await verifyAppleJwt(idToken);
    if (!claims) {
      return null;
    }

    return {
      email: claims.email?.toLowerCase() ?? null,
      appleId: claims.sub,
      name: claims.email?.split('@')[0],
    };
  } catch {
    return null;
  }
}

function getRefreshTokenMetadata(
  token: string,
): RefreshTokenPayload | null {
  const payload = verifyRefreshToken(token);
  if (!payload?.jti || typeof payload.exp !== 'number') {
    return null;
  }
  return payload;
}

async function isRefreshTokenRevoked(tokenId: string): Promise<boolean> {
  const rows = await db
    .select({ id: refreshTokenRevocations.id })
    .from(refreshTokenRevocations)
    .where(eq(refreshTokenRevocations.tokenId, tokenId))
    .limit(1);

  return rows.length > 0;
}

async function revokeRefreshToken(payload: RefreshTokenPayload): Promise<void> {
  if (!payload.jti || typeof payload.exp !== 'number') {
    return;
  }

  await db
    .insert(refreshTokenRevocations)
    .values({
      tokenId: payload.jti,
      userId: payload.userId,
      expiresAt: new Date(payload.exp * 1000),
    })
    .onConflictDoNothing();
}

export async function authRoutes(fastify: FastifyInstance) {
  const app = fastify.withTypeProvider<ZodTypeProvider>();

  /**
   * POST /auth/google - Login with Google
   */
  app.post(
    '/auth/google',
    {
      schema: {
        tags: ['Auth'],
        summary: 'Login with Google',
        description: 'Validate Google ID token, create or find user, return JWT tokens',
        body: z.object({
          idToken: z.string().min(1),
        }),
        response: {
          200: z.object({
            session: z.object({
              user: z.object({
                id: z.string(),
                username: z.string(),
                displayName: z.string(),
                profilePhotoUrl: z.string().nullable(),
                karma: z.number(),
                karmaRank: z.string(),
                createdAt: z.string(),
              }),
              accessToken: z.string(),
              refreshToken: z.string(),
              expiresAt: z.string(),
            }),
            isNewUser: z.boolean(),
          }),
          400: z.object({
            error: z.string(),
            message: z.string(),
          }),
          401: z.object({
            error: z.string(),
            message: z.string(),
          }),
        },
      },
    },
    async (request, reply) => {
      const { idToken } = request.body;

      // Validate Google token
      const googleUser = await validateGoogleToken(idToken);
      if (!googleUser) {
        return reply.status(401).send({
          error: 'INVALID_TOKEN',
          message: 'Invalid or expired Google ID token',
        });
      }

      // Check if user exists by Google ID or email
      let user = await db.query.users.findFirst({
        where: or(
          eq(users.googleId, googleUser.googleId),
          eq(users.email, googleUser.email)
        ),
      });

      let isNewUser = false;

      if (!user) {
        // Create new user
        isNewUser = true;
        user = await withGeneratedUniqueUsername(async (username) => {
          const displayName = googleUser.name || username;
          const [newUser] = await db
            .insert(users)
            .values({
              googleId: googleUser.googleId,
              email: googleUser.email,
              username,
              displayName,
            })
            .returning();

          return newUser;
        });
      } else if (!user.googleId) {
        // Link Google account to existing user
        await db
          .update(users)
          .set({ googleId: googleUser.googleId })
          .where(eq(users.id, user.id));
      }

      // Generate tokens
      const accessToken = generateAccessToken(fastify, user.id);
      const refreshToken = generateRefreshToken(user.id);
      const expiresAt = getAccessTokenExpiry();

      return {
        session: {
          user: {
            id: user.id,
            username: user.username,
            displayName: user.displayName || user.username,
            profilePhotoUrl: user.profilePhotoUrl,
            karma: user.karma,
            karmaRank: getKarmaRank(user.karma).title,
            createdAt: user.createdAt.toISOString(),
          },
          accessToken,
          refreshToken,
          expiresAt: expiresAt.toISOString(),
        },
        isNewUser,
      };
    }
  );

  /**
   * POST /auth/apple - Login with Apple
   */
  app.post(
    '/auth/apple',
    {
      schema: {
        tags: ['Auth'],
        summary: 'Login with Apple',
        description: 'Validate Apple ID token, create or find user, return JWT tokens',
        body: z.object({
          idToken: z.string().min(1),
        }),
        response: {
          200: z.object({
            session: z.object({
              user: z.object({
                id: z.string(),
                username: z.string(),
                displayName: z.string(),
                profilePhotoUrl: z.string().nullable(),
                karma: z.number(),
                karmaRank: z.string(),
                createdAt: z.string(),
              }),
              accessToken: z.string(),
              refreshToken: z.string(),
              expiresAt: z.string(),
            }),
            isNewUser: z.boolean(),
          }),
          400: z.object({
            error: z.string(),
            message: z.string(),
          }),
          401: z.object({
            error: z.string(),
            message: z.string(),
          }),
        },
      },
    },
    async (request, reply) => {
      const { idToken } = request.body;

      // Validate Apple token
      const appleUser = await validateAppleToken(idToken);
      if (!appleUser) {
        return reply.status(401).send({
          error: 'INVALID_TOKEN',
          message: 'Invalid or expired Apple ID token',
        });
      }

      // Check if user exists by Apple ID or email when Apple returned one.
      let user = await db.query.users.findFirst({
        where: appleUser.email
          ? or(eq(users.appleId, appleUser.appleId), eq(users.email, appleUser.email))
          : eq(users.appleId, appleUser.appleId),
      });

      let isNewUser = false;

      if (!user) {
        if (!appleUser.email) {
          return reply.status(400).send({
            error: 'EMAIL_REQUIRED',
            message:
              'Apple did not provide an email for this sign-in. Sign in with Apple again and share your email address.',
          });
        }
        const email = appleUser.email;

        // Create new user
        isNewUser = true;
        user = await withGeneratedUniqueUsername(async (username) => {
          const displayName = appleUser.name || username;
          const [newUser] = await db
            .insert(users)
            .values({
              appleId: appleUser.appleId,
              email,
              username,
              displayName,
            })
            .returning();

          return newUser;
        });
      } else if (!user.appleId) {
        // Link Apple account to existing user
        await db
          .update(users)
          .set({ appleId: appleUser.appleId })
          .where(eq(users.id, user.id));
      }

      // Generate tokens
      const accessToken = generateAccessToken(fastify, user.id);
      const refreshToken = generateRefreshToken(user.id);
      const expiresAt = getAccessTokenExpiry();

      return {
        session: {
          user: {
            id: user.id,
            username: user.username,
            displayName: user.displayName || user.username,
            profilePhotoUrl: user.profilePhotoUrl,
            karma: user.karma,
            karmaRank: getKarmaRank(user.karma).title,
            createdAt: user.createdAt.toISOString(),
          },
          accessToken,
          refreshToken,
          expiresAt: expiresAt.toISOString(),
        },
        isNewUser,
      };
    }
  );

  /**
   * POST /auth/refresh - Refresh access token
   */
  app.post(
    '/auth/refresh',
    {
      schema: {
        tags: ['Auth'],
        summary: 'Refresh access token',
        description: 'Exchange a refresh token for a new access token',
        body: refreshSchema,
        response: {
          200: z.object({
            accessToken: z.string(),
            expiresAt: z.string(),
          }),
          401: z.object({
            error: z.string(),
            message: z.string(),
          }),
        },
      },
    },
    async (request, reply) => {
      const { refreshToken } = request.body;

      const payload = getRefreshTokenMetadata(refreshToken);
      if (!payload) {
        return reply.status(401).send({
          error: 'INVALID_REFRESH_TOKEN',
          message: 'Invalid or expired refresh token',
        });
      }

      if (await isRefreshTokenRevoked(payload.jti)) {
        return reply.status(401).send({
          error: 'INVALID_REFRESH_TOKEN',
          message: 'Invalid or expired refresh token',
        });
      }

      // Verify user still exists
      const user = await db.query.users.findFirst({
        where: eq(users.id, payload.userId),
      });

      if (!user) {
        return reply.status(401).send({
          error: 'USER_NOT_FOUND',
          message: 'User no longer exists',
        });
      }

      // Generate new access token
      const accessToken = generateAccessToken(fastify, user.id);
      const expiresAt = getAccessTokenExpiry();

      return {
        accessToken,
        expiresAt: expiresAt.toISOString(),
      };
    }
  );

  /**
   * POST /auth/logout - Logout and invalidate refresh token
   */
  app.post(
    '/auth/logout',
    {
      schema: {
        tags: ['Auth'],
        summary: 'Logout',
        description: 'Invalidate refresh token (client should also clear tokens)',
        body: logoutSchema,
        response: {
          204: z.null(),
        },
      },
    },
    async (request, reply) => {
      const refreshToken = request.body.refreshToken;

      if (refreshToken) {
        const payload = getRefreshTokenMetadata(refreshToken);
        if (payload) {
          await revokeRefreshToken(payload);
        }
      }

      return reply.status(204).send(null);
    }
  );

  /**
   * GET /auth/me - Get current user profile
   */
  app.get(
    '/auth/me',
    {
      onRequest: [fastify.authenticate],
      schema: {
        tags: ['Auth'],
        summary: 'Get current user profile',
        description: 'Returns the profile of the currently authenticated user',
        response: {
          200: z.object({
            user: z.object({
              id: z.string(),
              username: z.string(),
              displayName: z.string(),
              profilePhotoUrl: z.string().nullable(),
              email: z.string(),
              karma: z.number(),
              karmaRank: z.string(),
              createdAt: z.string(),
            }),
          }),
          401: z.object({
            error: z.string(),
            message: z.string(),
          }),
        },
      },
    },
    async (request, reply) => {
      const userId = request.userId;

      if (!userId) {
        return reply.status(401).send({
          error: 'UNAUTHORIZED',
          message: 'Authentication required',
        });
      }

      const user = await db.query.users.findFirst({
        where: eq(users.id, userId),
      });

      if (!user) {
        return reply.status(401).send({
          error: 'USER_NOT_FOUND',
          message: 'User not found',
        });
      }

      return {
        user: {
          id: user.id,
          username: user.username,
          displayName: user.displayName || user.username,
          profilePhotoUrl: user.profilePhotoUrl,
          email: user.email,
          karma: user.karma,
          karmaRank: getKarmaRank(user.karma).title,
          createdAt: user.createdAt.toISOString(),
        },
      };
    }
  );
}